refactor: extract lotto service from App component
